chore: split key and path functions in separate files
